chore: share component render in scripts (#41603)
* chore: share component render in scripts

* chore: code clean
