+++ package.json
@@ -75,7 +75,7 @@
     "lint:demo": "eslint components/*/demo/*.md",
     "lint:deps": "antd-tools run deps-lint",
     "lint:md": "remark . -f -q",
-    "lint:style": "ts-node --project tsconfig.node.json scripts/check-cssinjs.js",
+    "lint:style": "ts-node --project tsconfig.node.json scripts/check-cssinjs",
     "lint:script": "eslint . --ext .js,.jsx,.ts,.tsx --cache",
     "pre-publish": "npm run test-all -- --skip-build",
     "prettier": "prettier -c --write **/* --cache",
