+++ scripts/check-cssinjs.js
@@ -1,15 +1,14 @@
 /* eslint-disable import/no-unresolved,no-console,global-require,import/no-dynamic-require */
 const chalk = require('chalk');
-const React = require('react');
-const ReactDOMServer = require('react-dom/server');
-const glob = require('glob');
-const path = require('path');
 const {
   StyleProvider,
   logicalPropertiesLinter,
   legacyNotSelectorLinter,
   parentSelectorLinter,
 } = require('@ant-design/cssinjs');
+const React = require('react');
+const ReactDOMServer = require('react-dom/server');
+const { generateCssinjs } = require('./generate-cssinjs');
 
 console.log(chalk.green(`🔥 Checking CSS-in-JS...`));
 
@@ -24,37 +23,17 @@ console.error = (msg) => {
   }
 };
 
-const EmptyElement = React.createElement('div');
-
-const styleFiles = glob.globSync(
-  path.join(
-    process.cwd(),
-    'components/!(version|config-provider|icon|auto-complete|col|row|time-picker)/style/index.?(ts|tsx)',
-  ),
-);
-
-styleFiles.forEach((file) => {
-  let useStyle = () => {};
-  if (file.includes('grid')) {
-    const { useColStyle, useRowStyle } = require(file);
-    useStyle = () => {
-      useRowStyle();
-      useColStyle();
-    };
-  } else {
-    useStyle = require(file).default;
-  }
-  const Component = () => {
-    useStyle('check');
-    return EmptyElement;
-  };
-  ReactDOMServer.renderToString(
-    React.createElement(
-      StyleProvider,
-      { linters: [logicalPropertiesLinter, legacyNotSelectorLinter, parentSelectorLinter] },
-      React.createElement(Component),
-    ),
-  );
+generateCssinjs({
+  key: 'check',
+  render: (Component) => {
+    ReactDOMServer.renderToString(
+      React.createElement(
+        StyleProvider,
+        { linters: [logicalPropertiesLinter, legacyNotSelectorLinter, parentSelectorLinter] },
+        React.createElement(Component),
+      ),
+    );
+  },
 });
 
 if (errorCount > 0) {
